Table

AgTables.AGTable — Type

AGTable

A base type that contains the necessary information to visualize a table.

See also: ag_table, ag_show, ag_save.

AgTables.ag_table — Function

ag_table(data::Vector{T}, column_defs::AbstractColumnDef...; kw...) -> AGTable

Creates a table using data, where elements can be NamedTuples, Dicts, or custom types. With column_defs you can configure parameters and formatting of columns.

Warning

Adding more than 100K data rows to a table may cause performance issues.

Keyword arguments

Name::Type	Default (Possible values)	Description

name::String	"AgTables ❤️ Julia"	Table name and tab title.

resize::Bool	true	Column width flexibility.

flex::Bool	true	Stretching columns to fit the available width.

header_height::Integer	40	Header height in px.

row_height::Integer	39	Row height in px.

column_filter::Bool	false	If a filter for column names should be used.

See also: ag_show, ag_save.

Example

using AgTables

struct Foo
    a::String
    b::String
    c::String
end

sample_data = [
    (a = "1", b = "2024-01-01", c = "tuple"),
    Dict("a" => "2", "b" => "2024-01-01", "d" => "dict"),
    Foo("3", "2024-01-01", "struct"),
]

table = ag_table(sample_data)

ag_save("table_example.html", table)

Panel

AgTables.AGPanel — Type

AGPanel

A type that allows you to combine multiple table for visualization.

See also: ag_panel, ag_show, ag_save.

AgTables.ag_panel — Function

ag_panel(tables::AGTable...; kw...) -> AGPanel

Creates a panel combining several tables.

Keyword arguments

Name::Type	Default/Posible values	Description

name::String	"AgTables ❤️ Julia"	Panel name (will be displayed in the browser tab title).

bottom::Bool	false	Move the table tabs to the bottom of the page.

See also: ag_show, ag_save.

Example

using AgTables

cars = [
    (id = 1, make = "Toyota", model = "Corolla", year = 2020),
    (id = 2, make = "Ford", model = "Mustang", year = 2019),
    (id = 3, make = "Honda", model = "Civic", year = 2021),
    (id = 4, make = "Chevrolet", model = "Implata", year = 2018),
]

customers = [
    (id = 1, name = "John", email = "john@example.com"),
    (id = 2, name = "Jane", email = "jane@example.com"),
    (id = 3, name = "Emily", email = "emily@example.com"),
    (id = 4, name = "Michael", email = "michael@example.com"),
]

rentals = [
    (rental_id = 1, car_id = 1, customer_id = 1, rental_date = "2024-06-01"),
    (rental_id = 2, car_id = 2, customer_id = 2, rental_date = "2024-06-05"),
    (rental_id = 3, car_id = 3, customer_id = 3, rental_date = "2024-06-07"),
    (rental_id = 4, car_id = 4, customer_id = 4, rental_date = "2024-06-09"),
]

table1 = ag_table(cars, name = "Cars")
table2 = ag_table(customers, name = "Customers")
table3 = ag_table(rentals, name = "Rentals")
panel = ag_panel(table1, table2, table3)

ag_save("panel_example.html", panel)

Columns

AgTables.AGColumns.AGThreshold — Type

AGThreshold(value; kw...)

Type used to visually color-code column values based on a threshold. Can be used for numeric, string, and time columns.

Keyword arguments

Name::Type	Default (Possible values)	Description

color_up::String	"#399746"	Color for values above the threshold.

color_down::String	"#ca3c32"	Color for values below the threshold.

String column

AgTables.AGColumns.AgStringColumnDef — Type

AgStringColumnDef(; kw...)

Type defining a string column and its display settings.

Keyword arguments

Name::Type	Default (Possible values)	Description

field_name::String	""	Column name in source data.

header_name::String	field_name	Override field_name as default with custom value.

visible::Bool	true	Column visibility.

filter::Bool	false	Enables a selector filter for column values.

default_sort::AG_SORT_MODES	nothing (AG_ASC, AG_DESC)	Initial sorting in the column (ascending or descending order).

str_format::String	"%s"	String template for formatting values. The %s is replaced with the value.

width::Integer	nothing	Column width.

cell_background::String	"#fff"	Cell background.

rect_background::String	"#fff"	Background color of the rectangle.

color::String	"#000"	Text color.

text_align::AG_TEXTALIGN_TYPES	AG_CENTER (AG_LEFT, AG_RIGHT)	Text alignment.

color_map::Dict{String,String}	Dict{String,String}()	Dictionary with colors of values (value => color).

threshold::AGThreshold	nothing	Settings for highlighting cells based on a threshold value.

filter_include::Vector{String}	String[]	Values that are active in the filter by default.

filter_exclude::Vector{String}	String[]	Values that are disabled in the filter by default.

Example

using AgTables

flights_sample_data = ag_flights_sample_data()

table = ag_table(
    flights_sample_data,
    AgStringColumnDef(
        field_name = "airline_name",
        header_name = "Airlines",
        cell_background = "#e8e8e8",
    ),
    AgNumberColumnDef(
        field_name = "flight_number",
        header_name = "Flight",
        color = "#00b4d8",
        width = 100,
    ),
    AgStringColumnDef(
        field_name = "destination",
        header_name = "Destination",
        filter = true,
        filter_exclude = ["Guatemala City", "Boston, MA"],
    ),
    AgStringColumnDef(
        field_name = "status",
        header_name = "Status",
        filter = true,
        rect_background = "#f0efeb",
        color_map = Dict(
            "In Air" => "#8ac926",
            "Canceled" => "#ff595e",
            "Scheduled" => "#999999",
        ),
        text_align = AG_CENTER,
    ),
)

ag_save("string_example.html", table)

Number column

AgTables.AGColumns.AGFormatter — Type

AGFormatter(; kw...)

Type used to configure the formatting of numeric column values.

Keyword arguments

Name::Type	Default (Possible values)	Description

short::Bool	false	Uses letter abbreviations to represent large numbers.

style::AG_STYLE_TYPES	AG_DECIMAL (AG_CURRENCY, AG_PERCENT)	The formatting style to use. Displays the corresponding currency or percent symbol next to the value.

currency::AG_CURRENCY_CODES	USD	The currency symbol used in formatting.

separator::Bool	false	Separates thousandths values with a comma.

minimum_fraction_digits::Integer	0	The minimum number of fraction digits to use.

maximum_fraction_digits::Integer	3	The maximum number of fraction digits to use.

AgTables.AGColumns.AgNumberColumnDef — Type

AgNumberColumnDef(; kw...)

Type defining a numeric column and its display settings.

Keyword arguments

Name::Type	Default (Possible values)	Description

field_name::String	""	Column name in source data.

header_name::String	field_name	Override field_name as default with custom value.

visible::Bool	true	Column visibility.

filter::Bool	false	Enables a filter for column values.

default_sort::AG_SORT_MODES	nothing (AG_ASC, AG_DESC)	Initial sorting in the column (ascending or descending order).

formatter::AGFormatter	nothing	Formatting of the numbers.

str_format::String	"%s"	String template for formatting values. The %s is replaced with the value.

width::Integer	nothing	Column width.

cell_background::String	"#fff"	Cell background.

rect_background::String	"#fff"	Background color of the rectangle.

color::String	"#000"	Text color.

text_align::AG_TEXTALIGN_TYPES	AG_CENTER (AG_LEFT, AG_RIGHT)	Text alignment.

color_map::Dict{Number,String}	Dict{Number,String}()	Dictionary with colors of values (value => color).

threshold::AGThreshold	nothing	Settings for highlighting cells based on a threshold value.

Example

using AgTables

currencies_sample_data = ag_currencies_sample_data()

table = ag_table(
    currencies_sample_data,
    AgStringColumnDef(
        field_name = "currency_code",
        header_name = "Currency",
        filter = true,
        text_align = AG_CENTER,
        rect_background = "#f0efeb",
        default_sort = AG_ASC,
    ),
    AgNumberColumnDef(
        field_name = "exchange_rate",
        header_name = "Exchange Rate",
        filter = true,
        formatter = AGFormatter(style = AG_PERCENT),
        threshold = AGThreshold(0.0, color_down = "#ff595e", color_up = "#8ac926"),
    ),
    AgNumberColumnDef(
        field_name = "market_cap",
        header_name = "Market Cap",
        filter = true,
        formatter = AGFormatter(
            style = AG_CURRENCY,
            currency = AgTables.USD,
            separator = true,
        ),
    ),
    AgNumberColumnDef(
        field_name = "daily_volume",
        header_name = "Daily Volume",
        filter = true,
        formatter = AGFormatter(short = true),
    ),
)

ag_save("number_example.html", table)

Time column

AgTables.AGColumns.AgTimeColumnDef — Type

AgTimeColumnDef(; kw...)

Type defining a time column and its display settings.

Keyword arguments

Name::Type	Default (Possible values)	Description

field_name::String	""	Column name in source data.

header_name::String	field_name	Override field_name as default with custom value.

visible::Bool	true	Column visibility.

filter::Bool	false	Enables a filter for column values.

default_sort::AG_SORT_MODES	nothing (AG_ASC, AG_DESC)	Initial sorting in the column (ascending or descending order).

date_formatter::AG_DATE_FORMATTER	AG_DATE_TIME (AG_TIME, AG_DATE)	Specify time display format.

str_format::String	"%s"	String template for formatting values. The %s is replaced with the value.

width::Integer	nothing	Column width.

cell_background::String	"#fff"	Cell background.

rect_background::String	"#fff"	Background color of the rectangle.

color::String	"#000"	Text color.

text_align::AG_TEXTALIGN_TYPES	AG_CENTER (AG_LEFT, AG_RIGHT)	Text alignment.

color_map::Dict{Union{Real,TimeType},String}	Dict{Real,String}()	Dictionary with colors of values (value => color).

threshold::AGThreshold	nothing	Settings for highlighting cells based on a threshold value.

Example

using AgTables

order_sample_data = ag_order_sample_data()

table = ag_table(
    order_sample_data,
    AgStringColumnDef(
        field_name = "symbol",
        text_align = AG_CENTER,
        width = 120,
    ),
    AgTimeColumnDef(
        field_name = "createdAt",
        filter = true,
        date_formatter = AG_DATE_TIME,
        text_align = AG_CENTER,
    ),
    AgStringColumnDef(
        field_name = "orderStatus",
        rect_background = "#f0efeb",
    ),
    AgStringColumnDef(field_name = "orderId", visible = false),
    AgStringColumnDef(field_name = "orderSide", visible = false),
    AgStringColumnDef(field_name = "orderType", visible = false),
    AgStringColumnDef(field_name = "timeInForce", visible = false),
    AgStringColumnDef(field_name = "origQty", visible = false),
    AgStringColumnDef(field_name = "executedQty", visible = false),
    AgStringColumnDef(field_name = "price", visible = false);
    column_filter = true,
    header_height = 60,
)

ag_save("time_example.html", table)